Create Student-in&out-info.c
This code is a simple program that allows the user to input information about a student, including their name, ID, and GPA, and then displays the entered information on the screen.

The code begins by including the standard input/output library (stdio.h). It then defines a structure called "students" using the typedef keyword. The structure contains three members: a character array called "name" to store the student's name, an integer variable called "ID" to store the student's ID, and a floating-point variable called "GPA" to store the student's grade point average.

The main function is the entry point of the program. It starts by declaring a variable of type "students" called "student". This variable will be used to store the information about the student.

The program prompts the user to enter the name of the student using the printf function and retrieves the input using the scanf function, which stores the input in the "name" member of the "student" structure.

Similarly, the program prompts the user to enter the ID and GPA of the student, and retrieves the input using scanf, storing the values in the respective members of the "student" structure.

Finally, the program uses printf to display the entered information on the screen, including the student's name, ID, and GPA, formatted using placeholders.

At the end of the main function, the code returns 0, indicating successful execution of the program.

Overall, this code provides a basic template for inputting and displaying information about a student using a structure in C programming language.