Create MultiBook-in&out-info.c
Outside the code:
This code snippet utilizes the C programming language to create a simple program for entering and displaying details of a book. It includes the standard input/output library (stdio.h), which provides functions for input and output operations.

Inside the code:

The code defines a structure called "Book" using the typedef keyword. The structure has three members: a character array called "title" to store the title of the book, a character array called "author" to store the name of the author, and an integer variable called "publicationYear" to store the year of publication.

The main function is the entry point of the program. It begins by declaring a variable of type "Book" called "book". This variable will be used to store the information about the book.

The program prompts the user to enter the title of the book using the printf function. It then uses the scanf function with a format specifier ("99[^\n]%*c") to read the input from the user and store it in the "title" member of the "book" structure. This format specifier allows input of up to 99 characters, excluding newline characters.

Similarly, the program prompts the user to enter the name of the author using printf, and retrieves the input using scanf with the same format specifier. The input is stored in the "author" member of the "book" structure.

The program prompts the user to enter the publication year of the book using printf. It then uses scanf with the "%d" format specifier to read the input as an integer and store it in the "publicationYear" member of the "book" structure.

After gathering the book details, the program prints a new line and displays the entered information on the screen using printf. The book details include the title, author, and publication year.

Finally, the code returns 0, indicating successful execution of the program.

Overall, this code allows the user to input details of a book, such as the title, author name, and publication year, and then displays the entered information on the screen using printf. It demonstrates the usage of structures in C programming for organizing related data.