Create Car-Rental-Catalog.c
This code implements a car rental system in C. It begins by defining a structure called "Car" that stores information about a car, including the brand, model, and year.

The main function is responsible for the execution of the program. It prompts the user to enter the number of cars they want to input by displaying the message "Enter the number of cars: ". The user's input is then stored in the variable "numCars".

Next, an array of structures called "cars" is declared, with the size determined by the value of "numCars". This array will store the details of each car.

A for loop is used to iterate over each car. Inside the loop, the program prompts the user to enter the details for each car, including the brand, model, and year. The input is read using the "scanf" function and stored in the corresponding members of the "cars" array.

After gathering all the car details, the program proceeds to display the car rental system. Another for loop is used to iterate over each car in the "cars" array. Inside the loop, the program prints the car number, brand, model, and year using the "printf" function.

Finally, the main function returns 0 to indicate successful program execution.

Overall, this code provides a basic car rental system that allows the user to input car details and displays them as part of the rental system.